pyproject.toml
[tool.poetry]
name = "imggen"
version = "0.1.0"
description = "image generator with stable-diffusion"
authors = ["saucehammer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
transformers = "^4.38.1"
python-slugify = "^8.0.4"
diffusers = {git = "https://github.com/huggingface/diffusers.git", rev = "main"}
ftfy = "^6.1.3"
scipy = "^1.12.0"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pillow = "^10.2.0"
black = "^24.2.0"
accelerate = "^0.27.2"
peft = "^0.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"